move header from primitive to component
