+++ ui/DesignSystem/Component/Transaction.svelte
@@ -1,10 +1,8 @@
 <script>
   import {
     costSummary,
-    formatMessage,
     formatStake,
     formatSubject,
-    SubjectType,
     PayerType,
   } from "../../src/transaction.ts";
 
@@ -17,6 +15,7 @@
   } from "../../DesignSystem/Primitive";
 
   import Rad from "./Rad.svelte";
+  import Header from "./Transaction/Header.svelte";
   import Row from "./Transaction/Row.svelte";
 
   export let transaction = null;
@@ -26,16 +25,6 @@
   let avatar;
 
   const subject = formatSubject(transaction.messages[0]);
-  const subjectAvatarShape = () => {
-    switch (subject.type) {
-      case SubjectType.User:
-      case SubjectType.Member:
-      case SubjectType.UserProject:
-        return "circle";
-      default:
-        return "square";
-    }
-  };
 
   const updateAvatar = async () => (avatar = await subject.avatarSource);
 
@@ -48,39 +37,7 @@
   $: updateAvatar();
 </script>
 
-<div dataCy="summary" variant={transaction.id ? 'top' : 'single'}>
-  <div
-    data-cy="message"
-    style="text-align:center; background-color: var(--color-foreground-level-1);
-    padding: 48px; border-radius: 4px; border: 1px solid
-    var(--color-foreground-level-2); margin-bottom:24px">
-    <div
-      style="background:var(--color-background); border-radius:28px;
-      height:56px; margin:0 auto 16px; padding:14px; width:56px; border:2px
-      solid var(--color-primary-level-1);">
-      <!-- TODO make the icon change based on transaction type -->
-      <Icon.Register style="fill: var(--color-primary)" />
-    </div>
-    <Title variant="big">{formatMessage(transaction.messages[0])}</Title>
-    <caption style="display: inline-flex; margin:16px 0;">
-      {#if avatar}
-        <Avatar
-          title={subject.name}
-          imageUrl={avatar.url}
-          avatarFallback={avatar.emoji && avatar}
-          variant={subjectAvatarShape()}
-          style="color: var(--color-foreground)"
-          dataCy="subject-avatar" />
-      {:else}
-        <Title>{subject.name}</Title>
-      {/if}
-    </caption>
-    <!-- TO DO Make the timestamp real -->
-    <Text variant="normal" style="color:var(--color-foreground-level-4);">
-      4 July 2020 16:20:21
-    </Text>
-  </div>
-</div>
+<Header {transaction} {avatar} {subject} />
 
 <Row dataCy="deposit" variant="top" style="">
   <div slot="left">
